feat: remove MarketWidget feature
