--- apps/ledger-live-desktop/src/mvvm/features/MarketPerformanceWidget/components/WidgetList.tsx
@@ -1,168 +1,0 @@
-import React, { useCallback } from "react";
-import { PropsBody, PropsBodyElem } from "../types";
-import { Flex, Text } from "@ledgerhq/react-ui";
-import styled from "@ledgerhq/react-ui/components/styled";
-import FormattedVal from "~/renderer/components/FormattedVal";
-import { MissingData } from "./MissingData";
-import { fontSizes } from "@ledgerhq/react-ui/styles/theme";
-import { counterValueCurrencySelector, localeSelector } from "~/renderer/reducers/settings";
-import { useSelector } from "LLD/hooks/redux";
-import counterValueFormatter from "@ledgerhq/live-common/market/utils/countervalueFormatter";
-import { useNavigate } from "react-router";
-import { track } from "~/renderer/analytics/segment";
-import { getChangePercentage } from "@ledgerhq/live-common/market/utils/index";
-import { CryptoIcon } from "@ledgerhq/crypto-icons";
-
-export function WidgetList({ data, order, range, top, enableNewFeature }: PropsBody) {
-  const noData = data.length === 0;
-
-  return (
-    <Flex flexDirection="column" flex={1} height={275}>
-      {noData ? (
-        <MissingData order={order} range={range} top={top} />
-      ) : (
-        data.map((elem, i) => (
-          <WidgetRow
-            key={i}
-            index={i + 1}
-            data={elem}
-            isFirst={i === 0}
-            range={range}
-            enableNewFeature={enableNewFeature}
-          />
-        ))
-      )}
-    </Flex>
-  );
-}
-
-function WidgetRow({ data, index, range, enableNewFeature }: PropsBodyElem) {
-  const counterValueCurrency = useSelector(counterValueCurrencySelector);
-  const locale = useSelector(localeSelector);
-  const navigate = useNavigate();
-
-  const onCurrencyClick = useCallback(() => {
-    track("widgetAsset_clicked", {
-      asset: data.name,
-      page: "Portfolio",
-    });
-
-    navigate(`/market/${data.id}`);
-  }, [data, navigate]);
-
-  return (
-    <MainContainer
-      justifyContent="space-between"
-      py="6px"
-      onClick={enableNewFeature ? onCurrencyClick : undefined}
-      featureFlagEnabled={enableNewFeature}
-      data-testid={`market-performance-widget-row-${index}`}
-    >
-      <Flex alignItems="center" flex={1}>
-        <Text color="neutral.c80" variant="h5Inter" mr={2}>
-          {index}
-        </Text>
-
-        <CryptoCurrencyIconWrapper>
-          {data.ledgerIds && data.ledgerIds.length > 0 && data.ticker ? (
-            <CryptoIcon ledgerId={data.ledgerIds[0]} ticker={data.ticker} size="32px" />
-          ) : (
-            <img width="32px" height="32px" src={data.image} alt={"currency logo"} />
-          )}
-        </CryptoCurrencyIconWrapper>
-
-        <Flex ml={2} overflow="hidden" flexDirection="column" flex={1} alignItems="left">
-          <EllipsisText
-            variant="paragraph"
-            fontWeight="semiBold"
-            color="neutral.c100"
-            fontSize={fontSizes.paragraph}
-          >
-            {data.name}
-          </EllipsisText>
-
-          <EllipsisText
-            variant="small"
-            color="neutral.c60"
-            fontWeight="medium"
-            fontSize={fontSizes.small}
-          >
-            {data.ticker.toUpperCase()}
-          </EllipsisText>
-        </Flex>
-      </Flex>
-
-      <Flex flexDirection="column">
-        <Text textAlign="right">
-          <FormattedVal
-            isPercent
-            isNegative
-            val={Number(parseFloat(String(getChangePercentage(data, range))).toFixed(2))}
-            inline
-            withIcon
-            style={{
-              fontSize: fontSizes.paragraph,
-              fontWeight: "medium",
-            }}
-          />
-        </Text>
-        <EllipsisText
-          variant="small"
-          textAlign="right"
-          color="neutral.c100"
-          fontWeight="medium"
-          fontSize={fontSizes.small}
-        >
-          {!data.price
-            ? "-"
-            : counterValueFormatter({
-                value: Number(parseFloat(String(data.price)).toFixed(data.price > 1 ? 2 : 6)),
-                currency: counterValueCurrency.ticker,
-                locale,
-              })}
-        </EllipsisText>
-      </Flex>
-    </MainContainer>
-  );
-}
-
-const MainContainer = styled(Flex)<{ featureFlagEnabled?: boolean }>`
-  transition:
-    background-color 0.35s ease,
-    padding 0.35s ease;
-  border-radius: 12px;
-
-  ${({ featureFlagEnabled, theme }) =>
-    featureFlagEnabled &&
-    `
-      &:hover {
-        transition-delay: 0.15s;
-        background-color: ${theme.colors.opacityDefault.c05};
-        padding: 6px 12px;
-        cursor: pointer;
-      }
-    `}
-`;
-
-const CryptoCurrencyIconWrapper = styled(Flex)<{
-  hasImage?: boolean;
-}>`
-  height: 32px;
-  width: 32px;
-  position: relative;
-  border-radius: 32px;
-  overflow: hidden;
-  background-color: ${p => (p.hasImage ? "transparent" : p.theme.colors.opacityDefault.c05)};
-  img {
-    object-fit: cover;
-  }
-  align-items: center;
-  justify-content: center;
-`;
-
-const EllipsisText = styled(Text)`
-  width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`;
