+++ apps/ledger-live-desktop/src/renderer/actions/constants.ts
@@ -3,7 +3,6 @@
 /** Settings --------- */
 export const PURGE_EXPIRED_ANONYMOUS_USER_NOTIFICATIONS =
   "settings/purgeExpiredAnonymousUserNotifications";
-export const TOGGLE_MARKET_WIDGET = "settings/toggleMarketWidget";
 export const TOGGLE_MEMOTAG_INFO = "settings/toggleShouldDisplayMemoTagInfo";
 export const TOGGLE_MEV = "settings/toggleMEV";
 export const UPDATE_ANONYMOUS_USER_NOTIFICATIONS = "settings/updateAnonymousUserNotifications";
