+++ apps/ledger-live-desktop/src/renderer/reducers/settings.ts
@@ -34,7 +34,6 @@ import regionsByKey from "~/renderer/screens/settings/sections/General/regions.j
 import { State } from ".";
 import {
   PURGE_EXPIRED_ANONYMOUS_USER_NOTIFICATIONS,
-  TOGGLE_MARKET_WIDGET,
   TOGGLE_MEMOTAG_INFO,
   TOGGLE_MEV,
   UPDATE_ANONYMOUS_USER_NOTIFICATIONS,
@@ -85,7 +84,6 @@ export type SettingsState = {
   accountsViewMode: "card" | "list";
   showAccountsHelperBanner: boolean;
   mevProtection: boolean;
-  marketPerformanceWidget: boolean;
   hideEmptyTokenAccounts: boolean;
   filterTokenOperationsZeroAmount: boolean;
   sidebarCollapsed: boolean;
@@ -183,7 +181,6 @@ export const INITIAL_STATE: SettingsState = {
   hasInstalledApps: true,
   lastSeenDevice: null,
   mevProtection: true,
-  marketPerformanceWidget: true,
   devicesModelList: [],
   lastSeenCustomImage: {
     size: 0,
@@ -281,7 +278,6 @@ type HandlersPayloads = {
   [PURGE_EXPIRED_ANONYMOUS_USER_NOTIFICATIONS]: { now: Date };
   [TOGGLE_MEV]: boolean;
   [TOGGLE_MEMOTAG_INFO]: boolean;
-  [TOGGLE_MARKET_WIDGET]: boolean;
   [UPDATE_ANONYMOUS_USER_NOTIFICATIONS]: {
     notifications: Record<string, number>;
   };
@@ -464,10 +460,6 @@ const handlers: SettingsHandlers = {
   [TOGGLE_MEV]: (state: SettingsState, { payload }) => ({
     ...state,
     mevProtection: payload,
-  }),
-  [TOGGLE_MARKET_WIDGET]: (state: SettingsState, { payload }) => ({
-    ...state,
-    marketPerformanceWidget: payload,
   }),
   [TOGGLE_MEMOTAG_INFO]: (state: SettingsState, { payload }) => ({
     ...state,
@@ -771,8 +763,6 @@ export const hasBeenRedirectedToPostOnboardingSelector = (state: State) =>
 export const lastOnboardedDeviceSelector = (state: State) => state.settings.lastOnboardedDevice;
 
 export const mevProtectionSelector = (state: State) => state.settings.mevProtection;
-export const marketPerformanceWidgetSelector = (state: State) =>
-  state.settings.marketPerformanceWidget;
 export const alwaysShowMemoTagInfoSelector = (state: State) => state.settings.alwaysShowMemoTagInfo;
 export const anonymousUserNotificationsSelector = (state: State) =>
   state.settings.anonymousUserNotifications;
