+++ apps/ledger-live-desktop/static/i18n/en/app.json
@@ -1477,34 +1477,6 @@
         "primaryCta": "Connect device",
         "secondaryCta": "Maybe later"
       }
-    },
-    "marketPerformanceWidget": {
-      "title": "{{time}} trend",
-      "buy": "Buy",
-      "swap": "Swap",
-      "gainers": "Gainers",
-      "losers": "Losers",
-      "error": {
-        "title": "Sorry, we’re unable to load the trend",
-        "description": "Please try again later"
-      },
-      "missingData": {
-        "gainers": {
-          "title": "No assets among the top {{count}} experienced gains in the last {{trend}}",
-          "description": "Consider viewing top losers or come back later"
-        },
-        "losers": {
-          "title": "No assets among the top {{count}} experienced losses in the last {{trend}}",
-          "description": "Consider viewing top gainers or come back later"
-        }
-      },
-      "trend": {
-        "day": "24 hours",
-        "week": "7 days",
-        "month": "30 days",
-        "year": "year",
-        "all": "year"
-      }
     }
   },
   "currentAddress": {
