+++ apps/ledger-live-desktop/tests/specs/general/portfolio.spec.ts
@@ -1,6 +1,5 @@
 import test from "../../fixtures/common";
 import { expect } from "@playwright/test";
-import { Layout } from "../../component/layout.component";
 import { PortfolioPage } from "../../page/portfolio.page";
 
 test.use({
@@ -13,7 +12,6 @@ test.use({
 });
 
 test("Portfolio @smoke", async ({ page }) => {
-  const layout = new Layout(page);
   const portfolioPage = new PortfolioPage(page);
 
   await test.step("load portfolio", async () => {
@@ -21,9 +19,7 @@ test("Portfolio @smoke", async ({ page }) => {
     await portfolioPage.checkBuySellButtonVisibility();
     await portfolioPage.checkSwapButtonVisibility();
     await portfolioPage.checkStakeButtonVisibility();
-    await expect.soft(page).toHaveScreenshot("portfolio.png", {
-      mask: [layout.marketPerformanceWidget],
-    });
+    await expect.soft(page).toHaveScreenshot("portfolio.png");
   });
 
   await test.step(`scroll to operations`, async () => {
