+++ e2e/desktop/tests/page/portfolio.page.ts
@@ -9,11 +9,7 @@ export class PortfolioPage extends AppPage {
   private stakeEntryButton = this.page.getByTestId("stake-entry-button");
   private chart = this.page.getByTestId("chart-container");
   private operationList = this.page.locator("#operation-list");
-  private marketPerformanceWidget = this.page.getByTestId("market-performance-widget");
-  private swapButton = this.marketPerformanceWidget.getByRole("button", { name: "Swap" });
-  private buyButton = this.marketPerformanceWidget.getByRole("button", { name: "Buy" });
   private assetAllocationTitle = this.page.getByText("Asset allocation");
-  private trendTitle = this.marketPerformanceWidget.getByText("1W trend");
   private assetRowElements = this.page.locator("[data-testid^='asset-row-']");
   private showAllButton = this.page.getByText("Show all");
   private showMoreButton = this.page.getByText("Show more");
@@ -63,20 +59,6 @@ export class PortfolioPage extends AppPage {
   @step("Check chart visibility")
   async checkChartVisibility() {
     await expect(this.chart).toBeVisible();
-  }
-
-  @step("Check market performance trend visibility")
-  async checkMarketPerformanceTrendVisibility() {
-    await expect(this.marketPerformanceWidget).toBeVisible();
-    await expect(this.trendTitle).toBeVisible();
-    await expect(this.buyButton).toBeVisible();
-    await expect(this.swapButton).toBeVisible();
-    await this.bestPerformersButton.click();
-    await expect(this.performerArrowDirection("up")).toBeVisible();
-    const bestPerformer = await this.rowOnePerformer.innerText();
-    await this.worstPerformersButton.click();
-    await expect(this.performerArrowDirection("down")).toBeVisible();
-    await expect(this.rowOnePerformer).not.toContainText(bestPerformer ?? "");
   }
 
   @step("Check asset allocation section")
