+++ e2e/desktop/tests/userdata/swap-history.json
@@ -34,7 +34,6 @@
       "hasInstalledApps": true,
       "lastSeenDevice": null,
       "mevProtection": true,
-      "marketPerformanceWidget": true,
       "devicesModelList": [],
       "lastSeenCustomImage": { "size": 0, "hash": "" },
       "latestFirmware": null,
