+++ e2e/mobile/userdata/speculos-subAccount.json
@@ -31,7 +31,6 @@
       "hasInstalledApps": true,
       "lastSeenDevice": null,
       "mevProtection": true,
-      "marketPerformanceWidget": true,
       "devicesModelList": [],
       "lastSeenCustomImage": {
         "size": 0,
