+++ e2e/mobile/userdata/speculos-tests-app.json
@@ -55,7 +55,6 @@
         "apps": []
       },
       "mevProtection": true,
-      "marketPerformanceWidget": true,
       "devicesModelList": ["europa", "stax"],
       "lastSeenCustomImage": { "size": 0, "hash": "" },
       "latestFirmware": null,
