+++ e2e/mobile/userdata/swap-history.json
@@ -38,7 +38,6 @@
       "hasInstalledApps": true,
       "lastSeenDevice": null,
       "mevProtection": true,
-      "marketPerformanceWidget": true,
       "devicesModelList": [],
       "lastSeenCustomImage": { "size": 0, "hash": "" },
       "latestFirmware": null,
